+++ Misc/NEWS.d/next/Windows/2019-02-11-14-53-17.bpo-35947.9vI4hP.rst
@@ -1,0 +1,2 @@
+Added current version of libffi to cpython-source-deps. 
+Change _ctypes to use current version of libffi on Windows.
